my_music_recommender/requirements.txt
# requirements.txt
Flask>=2.0
numpy
librosa
soundfile
scikit-learn
requests
spotipy
openai
faiss-cpu
gunicorn
markdown 